rearchitecture render article methods from articleRenderers.ts file
