+++ src/util/index.ts
@@ -1,7 +1,6 @@
 export * from './articleListMainPage.js'
 export * from './misc.js'
 export * from './dump.js'
-export * from './articleRenderers.js'
 export * from './const.js'
 export * from './mw-api.js'
 export * from './metaData.js'
