+++ test/unit/saveArticles.test.ts
@@ -5,8 +5,9 @@ import { setupScrapeClasses } from '../util.js'
 import { saveArticles, getModuleDependencies, applyOtherTreatments } from '../../src/util/saveArticles.js'
 import { ZimArticle } from '@openzim/libzim'
 import { Dump } from '../../src/Dump.js'
-import { mwRetToArticleDetail, renderDesktopArticle, DELETED_ARTICLE_ERROR } from '../../src/util/index.js'
+import { mwRetToArticleDetail, DELETED_ARTICLE_ERROR } from '../../src/util/index.js'
 import { jest } from '@jest/globals'
+import articleRenderer from '../../src/util/renderers/article.renderer.js'
 
 jest.setTimeout(40000)
 
@@ -193,7 +194,7 @@ describe('saveArticles', () => {
       visualeditor: { oldid: 0 },
     }
     // Throwing error if article is deleted
-    expect(() => renderDesktopArticle(articleJsonObject, 'deletedArticle', { title: 'deletedArticle' })).toThrow(new Error(DELETED_ARTICLE_ERROR))
+    expect(() => articleRenderer.renderDesktopArticle(articleJsonObject, 'deletedArticle', { title: 'deletedArticle' })).toThrow(new Error(DELETED_ARTICLE_ERROR))
   })
 
   test('Load inline js from HTML', async () => {
